Rename move from jamendo to marionette
Initially, I wanted to make a proof of concept based on the idea that one logs
in to one service, and only then shows useful stuff (and I chose Jamendo). But
I've changed my mind, and I'll fake the login completely. So now is a proof of
concept on how to do that with Marionette.